agent/requirements.txt
prometheus-client==0.14.1
flask==1.0.3
psutil==5.6.6
pyyaml==6.0
psutil==5.9.1
web3==5.29.2
